+++ src/main/java/com/blog/lastdance/entity/Comments.java
@@ -1,0 +1,34 @@
+package com.blog.lastdance.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+import java.time.*;
+
+@Entity
+@Getter
+@Setter
+@Table(name = "comments")
+public class Comments {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    private Long id;
+
+    @Column(name = "content")
+    private String content;
+
+    @Column(name = "created_at")
+    private LocalDateTime createdAt;
+
+    @Column(name = "updated_at")
+    private LocalDateTime updatedAt;
+
+    @Column(name = "member_id")
+    private Integer memberId;
+
+    @Column(name = "post_id")
+    private Integer postId;
+
+}
