+++ src/main/java/com/blog/lastdance/entity/Posts.java
@@ -1,0 +1,51 @@
+package com.blog.lastdance.entity;
+
+import com.blog.lastdance.dto.postdto.PostsRequestDTO;
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@Setter
+@Table(name = "posts")
+public class Posts {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    private Long id;
+
+    @Column(name = "title")
+    private String title;
+
+    @Column(name = "content")
+    private String content;
+
+    @Column(name = "great")
+    private Integer great;
+
+    @Column(name = "comment_count")
+    private Integer commentCount;
+
+    @Column(name = "created_at")
+    private LocalDateTime createdAt;
+
+    @Column(name = "updated_at")
+    private LocalDateTime updatedAt;
+
+    @Column(name = "author_Id")
+    private Long authorId;
+
+//    @ManyToOne
+//    private Members member;
+
+    public void update(PostsRequestDTO dto) {
+        this.title = dto.getTitle();
+        this.content = dto.getContent();
+        this.updatedAt = LocalDateTime.now();
+    }
+
+}
