+++ src/main/java/com/blog/lastdance/exception/CustomException.java
@@ -1,0 +1,15 @@
+package com.blog.lastdance.exception;
+
+import lombok.Getter;
+
+@Getter
+public class CustomException extends RuntimeException {
+
+    private final int status;
+
+    public CustomException(String message, int status) {
+        super(message);
+        this.status = status;
+    }
+
+}
