+++ src/main/java/com/blog/lastdance/exception/GlobalExceptionHandler.java
@@ -1,0 +1,34 @@
+package com.blog.lastdance.exception;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import java.util.Map;
+
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+
+    @ExceptionHandler(CustomException.class)
+    public ResponseEntity<?> handleCustomException(CustomException e) {
+        return ResponseEntity
+                .status(e.getStatus())
+                .body(Map.of(
+                        "error", e.getMessage()
+                ));
+    }
+
+
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<?> handleException(Exception e) {
+        e.printStackTrace();
+        return ResponseEntity
+                .status(500)
+                .body(Map.of(
+                        "error", "서버 오류가 발생했습니다.",
+                        "details", e.getMessage()
+                ));
+    }
+
+
+}
