+++ src/main/java/com/blog/lastdance/service/CustomMemberDetailService.java
@@ -1,0 +1,24 @@
+package com.blog.lastdance.service;
+
+import com.blog.lastdance.detail.CustomMemberDetails;
+import com.blog.lastdance.entity.Members;
+import com.blog.lastdance.repository.MembersRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class CustomMemberDetailService implements UserDetailsService {
+
+    private final MembersRepository membersRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String userId) throws UsernameNotFoundException {
+        Members members = membersRepository.findByUserId(userId)
+                .orElseThrow(() -> new UsernameNotFoundException("사용자를 찾을 수 없습니다: " + userId));
+        return new CustomMemberDetails(members);
+    }
+}
