--- src/main/java/com/example/demo/DemoApplication.java
@@ -1,18 +1,0 @@
-package com.example.demo;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DemoApplication {
-
-	public static void main(String[] args) {
-		try {
-			SpringApplication.run(DemoApplication.class, args);
-		}
-		catch (Exception e) {
-			e.printStackTrace();
-		}
-	}
-
-}
