--- src/main/java/com/example/demo/dto/LoginDTO.java
@@ -1,21 +1,0 @@
-package com.example.demo.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@ToString
-public class LoginDTO {
-
-    private String userId;
-    private String password;
-
-    public UsernamePasswordAuthenticationToken toAuthToken() {
-        return new UsernamePasswordAuthenticationToken(userId, password);
-    }
-}
