--- src/main/java/com/example/demo/entity/Pictures.java
@@ -1,24 +1,0 @@
-package com.example.demo.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-@Getter
-@Setter
-@Table(name = "pictures")
-public class Pictures {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Long id;
-
-    @Column(name = "picture")
-    private String picture;
-
-    @Column(name = "post_id")
-    private Long postId;
-
-}
