--- src/main/java/com/example/demo/entity/Posts.java
@@ -1,51 +1,0 @@
-package com.example.demo.entity;
-
-import com.example.demo.dto.PostsDTO;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Getter
-@Setter
-@Table(name = "posts")
-public class Posts {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Long id;
-
-    @Column(name = "title")
-    private String title;
-
-    @Column(name = "content")
-    private String content;
-
-    @Column(name = "great")
-    private Integer great;
-
-    @Column(name = "comment_count")
-    private Integer commentCount;
-
-    @Column(name = "created_at")
-    private LocalDateTime createdAt;
-
-    @Column(name = "updated_at")
-    private LocalDateTime updatedAt;
-
-    @Column(name = "author_Id")
-    private Long authorId;
-
-//    @ManyToOne
-//    private Members member;
-
-    public void update(PostsDTO dto) {
-        this.title = dto.getTitle();
-        this.content = dto.getContent();
-        this.updatedAt = LocalDateTime.now();
-    }
-
-}
