--- src/main/java/com/example/demo/exception/CustomException.java
@@ -1,15 +1,0 @@
-package com.example.demo.exception;
-
-import lombok.Getter;
-
-@Getter
-public class CustomException extends RuntimeException {
-
-    private final int status;
-
-    public CustomException(String message, int status) {
-        super(message);
-        this.status = status;
-    }
-
-}
