--- src/main/java/com/example/demo/exception/GlobalExceptionHandler.java
@@ -1,34 +1,0 @@
-package com.example.demo.exception;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.util.Map;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(CustomException.class)
-    public ResponseEntity<?> handleCustomException(CustomException e) {
-        return ResponseEntity
-                .status(e.getStatus())
-                .body(Map.of(
-                        "error", e.getMessage()
-                ));
-    }
-
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<?> handleException(Exception e) {
-        e.printStackTrace();
-        return ResponseEntity
-                .status(500)
-                .body(Map.of(
-                        "error", "서버 오류가 발생했습니다.",
-                        "details", e.getMessage()
-                ));
-    }
-
-
-}
