--- src/main/java/com/example/demo/repository/MembersRepository.java
@@ -1,11 +1,0 @@
-package com.example.demo.repository;
-
-import com.example.demo.entity.Members;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface MembersRepository extends JpaRepository<Members, Integer> {
-    Optional<Members> findByUserId(String userId);
-    Optional<Members> findByNickname(String nickName);
-}
