--- src/main/java/com/example/demo/service/CustomUserDetailService.java
@@ -1,24 +1,0 @@
-package com.example.demo.service;
-
-import com.example.demo.detail.CustomMemberDetails;
-import com.example.demo.entity.Members;
-import com.example.demo.repository.MembersRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class CustomUserDetailService implements UserDetailsService {
-
-    private final MembersRepository membersRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String userId) throws UsernameNotFoundException {
-        Members members = membersRepository.findByUserId(userId)
-                .orElseThrow(() -> new UsernameNotFoundException("사용자를 찾을 수 없습니다: " + userId));
-        return new CustomMemberDetails(members);
-    }
-}
