+++ src/test/java/com/blog/lastdance/LastDanceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.blog.lastdance;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class LastDanceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
